Increase MOSFET gate resistors to 180 ohms
The new value was calculated by considering the instant at which the
microcontroller pin goes high. The microcontroller is driving 3.3V at
one end of the resistor, while the voltage at the MOSFET gate changes
relatively slowly due to gate capacitance and so is assumed to still be
0V. We aim for a max current of 18mA (below the microcontroller's limit
of 20mA), giving a resistance of 3.3V/18mA = 183 ohms.

The previous value of 100 ohms was pulled off some forum thread somewhere.