schematic: change feedback threshold voltage divider resistances to reduce the number of different types of resistor needed by the system
Now, when the threshold adjust potentiometer is at its central position,
SYNC_THRESH is 0.16V away from the nominal 8.6V (as compared to 0.07V before
this change). This is deemed tolerable since the voltage can be adjusted
back to 8.6V using the potentiometer.

Previously the 8.2k and 11k resistors were the only ones of their
kinds used in the design. Additionally, the 11k was the only non-E12
resistor in the design. Now, only the 4.7k resistor is unique; 6.8k
is shared with R13 (LCD backlight current limiting resistor).