schematic: change debouncing filter resistors to reduce the number of different types of resistor in the system
The particular value of resistor for the debouncing filter doesn't really
matter too much. It was easier to change all of these from 33k to 68k,
rather than to change the one previously existing 68k: R44, which is part
of the voltage divider for SYNC and can't be changed without also
changing R47.