schematic: add an x-voltage-rating property to all capacitors
x-min-voltage-rating is set to the minimum voltage rating necessary
for the circuit to operate, as determined by the context the
component is being used in.

x-voltage-rating is set to the voltage rating of the physical component
that will be used here, and must be greater than or equal to x-min-voltage-rating.

This facilitates BOM generation by allowing capacitors with different
minimum voltage ratings to be grouped together.