firmware/BlockStorage: use standard boot jump code in the header
Windows seems to treat this as a fixed signature rather than as
arbitrary x86 code (or at least it doesn't like what I had in there
before). Changing to the commonly-accepted EBh 3Ch 90h (a jump to
just past the end of the FAT header) results in the device working
fine when plugged into a Windows 10 computer now.

A side effect is that now weirdness will occur if (for some reason)
you boot from the device. It would be ideal to print a message and
halt.