firmware: print a useful message if the user tries to boot from the device
Since the device is a USB Mass Storage Device, the BIOS will often
try to boot from it if it is plugged in while the computer is powered
on. In the event that this happens, it's preferable to display a message
to the user and halt than to let undefined behaviour occur.